+++ src/felupe/quadrature/_base.py
@@ -30,21 +30,56 @@ class Scheme:
 
         self.npoints, self.dim = self.points.shape
 
-    def plot(self, plotter, **kwargs):
-        """Plot the quadrature points, scaled by their weights, into a given PyVista
-        plotter.
+    def plot(self, plotter=None, add_axes=True, point_size=100, **kwargs):
+        """Plot the quadrature points, scaled by their weights, into a (optionally
+        provided) PyVista plotter.
 
         See Also
         --------
         felupe.Scene.plot: Plot method of a scene.
         """
 
+        if plotter is None:
+            from ..mesh import Mesh
+
+            mesh = Mesh(self.points, np.zeros((0, 2), dtype=int), "line")
+            plotter = mesh.plot(**kwargs)
+
         for weight, point in zip(self.weights, self.points):
+            scalars = weight / self.weights.max()
             plotter.add_points(
                 points=np.pad([point], ((0, 0), (0, 3 - self.points.shape[1]))),
-                point_size=100 * weight,
+                point_size=point_size * scalars,
                 render_points_as_spheres=True,
+                opacity=0.8,
                 color="grey",
             )
 
+        if add_axes:
+            plotter.add_axes(xlabel="r", ylabel="s", zlabel="t")
+
         return plotter
+
+    def screenshot(
+        self,
+        *args,
+        filename=None,
+        transparent_background=None,
+        scale=None,
+        **kwargs,
+    ):
+        """Take a screenshot of the quadrature.
+
+        See Also
+        --------
+        pyvista.Plotter.screenshot: Take a screenshot of a PyVista plotter.
+        """
+
+        if filename is None:
+            filename = "quadrature.png"
+
+        return self.plot(*args, off_screen=True, **kwargs).screenshot(
+            filename=filename,
+            transparent_background=transparent_background,
+            scale=scale,
+        )