+++ tests/test_quadrature.py
@@ -49,6 +49,8 @@ def test_gausslegendre():
         assert np.isclose(q23.weights.sum(), 8)
 
     q23 = fe.GaussLegendre(order=2, dim=3)
+    q23.plot(off_screen=True)
+
     assert q23.inv().points.shape == q23.points.shape
     assert np.allclose(q23.weights, q23.inv().weights)